+++ undine/api/database/base_client.py
@@ -1,10 +1,11 @@
-from undine.utils.exception import VirtualMethodException
+from undine.database.rabbitmq import RabbitMQConnector
+from undine.utils.exception import UndineException, VirtualMethodException
 
 import json
 import uuid
 
 
-class ClientBase:
+class BaseClient:
     def __init__(self, publish_task=None):
         if publish_task:
             self.publish_task = publish_task
@@ -85,3 +86,58 @@ class ClientBase:
 
     def _insert_task(self, _task):
         raise VirtualMethodException(self.__class__, '_insert_task')
+
+
+class BaseNetworkClient(BaseClient):
+    def __init__(self, task_queue):
+        BaseClient.__init__(self, self.__default_publish_task_with_mid)
+
+        if task_queue is None:
+            raise UndineException('Missing RabbitMQ option field (task_queue)')
+
+        self._queue = RabbitMQConnector(task_queue, consumer=False)
+
+    #
+    # Private methods
+    #
+    def __default_publish_task_with_mid(self, name, cid, iid, wid, mid, report):
+        task = {
+            'tid': self._get_uuid(),
+            'name': name,
+            'cid': cid,
+            'iid': iid,
+            'wid': wid,
+            'mid': mid,
+            'reportable': report
+        }
+
+        # Insert into remote host
+        self._insert_task(task)
+
+        # Insert useful task information into rabbitmq task queue
+        del task['name'], task['mid'], task['reportable']
+
+        self._queue.publish(json.dumps(task))
+
+        return task['tid']
+
+    #
+    # Protected inherited method
+    #
+    def _insert_mission(self, _mission):
+        raise VirtualMethodException(self.__class__, '_insert_mission')
+
+    #
+    # Public methods
+    #
+    def publish_mission(self, name, email, description):
+        mission = {
+            'mid': self._get_uuid(),
+            'name': name,
+            'email': email,
+            'description': description
+        }
+
+        self._insert_mission(mission)
+
+        return mission['mid']