+++ undine/client/database/mariadb.py
@@ -1,20 +1,20 @@
 from collections import namedtuple
-from undine.client.database.client_base import ClientBase
+from undine.client.database.base_client import BaseClient
 from undine.database.mariadb import MariaDbConnector
 from undine.utils.exception import UndineException
 
 
-class MariaDbClient(ClientBase):
+class MariaDbClient(BaseClient):
     _QUERY = {
         'mission_list': '''
           SELECT mid, name, email,
                  ready, issued, done, canceled, failed, 
-                 DATE_FORMAT(issued_at, '%Y-%m-%d %T')
+                 DATE_FORMAT(issued_at, '%Y-%m-%d %T') AS issued
             FROM mission_dashboard {where}
         ORDER BY complete, issued_at DESC
         ''',
         'mission_info': '''
-          SELECT HEX(mid), name, email, description,
+          SELECT HEX(mid) AS mid, name, email, description,
                  DATE_FORMAT(issued, '%Y-%m-%d %T')
             FROM mission {where}
         ORDER BY issued ASC
@@ -64,6 +64,37 @@ class MariaDbClient(ClientBase):
                  DATE_FORMAT(issued, '%Y-%m-%d %T') AS issued
             FROM worker {where}
         ORDER BY issued
+        ''',
+        'host_list': '''
+          SELECT name, ip, issued, canceled, failed,
+                 registered, logged_in, logged_out, state
+            FROM host_list
+        ''',
+        'reset_list': '''
+          SELECT HEX(tid), HEX(cid), HEX(iid), HEX(wid) 
+            FROM task 
+           WHERE tid in (SELECT tid FROM task {where})
+        ''',
+        'trash_result': '''
+          INSERT INTO trash (tid, generated, category, content)
+          SELECT tid, reported, 'result', content FROM result
+           WHERE tid = UNHEX(%(tid)s)
+        ''',
+        'trash_error': '''
+          INSERT INTO trash (tid, generated, category, content)
+          SELECT tid, informed, 'error', message FROM error
+           WHERE tid = UNHEX(%(tid)s)
+        ''',
+        'delete_result': '''
+          DELETE FROM result WHERE tid = UNHEX(%(tid)s)
+        ''',
+        'delete_error': '''
+          DELETE FROM error WHERE tid = UNHEX(%(tid)s)
+        ''',
+        'reset_task': '''
+          UPDATE task
+             SET state = 'R', host = NULL, ip = NULL
+           WHERE tid = UNHEX(%(tid)s)
         '''
     }
 
@@ -85,7 +116,8 @@ class MariaDbClient(ClientBase):
     # Constructor & Destructor
     #
     def __init__(self, config):
-        self._mariadb = MariaDbConnector(config)
+        BaseClient.__init__(self, config)
+        self._db = MariaDbConnector(self.db_config)
 
     #
     # Private methods
@@ -94,20 +126,23 @@ class MariaDbClient(ClientBase):
     def _value(form, value):
         return value if not form else form.format(value)
 
-    def _build_query(self, template, **kwargs):
-        where = list()
+    @staticmethod
+    def _where(condition):
+        return 'WHERE ' + ' AND '.join(condition) if condition else ''
+
+    def _build_query(self, query, **kwargs):
+        condition = list()
         params = dict()
 
-        for key, value in kwargs.items():
-            if key in self._WHERE_CLAUSE:
-                where.append(self._WHERE_CLAUSE[key].clause)
-                params[key] = self._value(self._WHERE_CLAUSE[key].format, value)
+        for k, v in kwargs.items():
+            if k in self._WHERE_CLAUSE:
+                condition.append(self._WHERE_CLAUSE[k].clause)
+                params[k] = self._value(self._WHERE_CLAUSE[k].format, v)
 
-        where = 'WHERE ' + ' AND '.join(where) if where else ''
+        return {'query': query.format(self._where(condition)), 'params': params}
 
-        query = template.format(where=where)
-
-        return {'query': query, 'params': params}
+    def _build_sql_item(self, template, **kwargs):
+        return self._db.sql_item(**self._build_query(template, **kwargs))
 
     #
     # Inherited methods
@@ -116,59 +151,73 @@ class MariaDbClient(ClientBase):
         where = 'WHERE complete = FALSE' if not list_all else ''
         query = self._QUERY['mission_list'].format(where=where)
 
-        return self._mariadb.fetch_all_tuples(query)
+        return self._db.fetch_all_tuples(query)
 
     def mission_info(self, **kwargs):
         if 'mid' in kwargs and len(kwargs['mid']) != 32:
             raise UndineException('MID value must feat to uuid length.')
 
         query_set = self._build_query(self._QUERY['mission_info'], **kwargs)
-        return self._mariadb.fetch_all_tuples(**query_set)
+        return self._db.fetch_all_tuples(**query_set)
 
     def task_list(self, **kwargs):
         if 'mid' not in kwargs or len(kwargs['mid']) != 32:
             raise UndineException('MID value must feat to uuid length.')
 
         query_set = self._build_query(self._QUERY['task_list'], **kwargs)
-        return self._mariadb.fetch_all_tuples(**query_set)
+        return self._db.fetch_all_tuples(**query_set)
 
     def task_info(self, tid):
         if len(tid) != 32:
             raise UndineException('TID value must feat to uuid length.')
 
-        item = {'tid': tid}
-        return self._mariadb.fetch_a_tuple(self._QUERY['task_info'], item)
+        return self._db.fetch_a_tuple(self._QUERY['task_info'], {'tid': tid})
 
     def config_info(self, cid):
         if len(cid) != 32:
             raise UndineException('CID value must feat to uuid length.')
 
-        item = {'cid': cid}
-        return self._mariadb.fetch_a_tuple(self._QUERY['config_info'], item)
+        return self._db.fetch_a_tuple(self._QUERY['config_info'], {'cid': cid})
 
     def input_info(self, iid):
         if len(iid) != 32:
             raise UndineException('IID value must feat to uuid length.')
 
-        query_set = self._build_query(self._QUERY['input_info'],
-                                      **{'iid': iid})
-        return self._mariadb.fetch_a_tuple(**query_set)
+        query_set = self._build_query(self._QUERY['input_info'], iid=iid)
+        return self._db.fetch_a_tuple(**query_set)
 
     def input_list(self):
         query_set = self._build_query(self._QUERY['input_info'])
 
-        print(query_set['query'])
-
-        return self._mariadb.fetch_all_tuples(**query_set)
+        return self._db.fetch_all_tuples(**query_set)
 
     def worker_info(self, wid):
         if len(wid) != 32:
             raise UndineException('WID value must feat to uuid length.')
 
-        query_set = self._build_query(self._QUERY['worker_info'],
-                                      **{'wid': wid})
-        return self._mariadb.fetch_a_tuple(**query_set)
+        query_set = self._build_query(self._QUERY['worker_info'], wid=wid)
+        return self._db.fetch_a_tuple(**query_set)
 
     def worker_list(self):
         query_set = self._build_query(self._QUERY['worker_info'])
-        return self._mariadb.fetch_all_tuples(**query_set)
+        return self._db.fetch_all_tuples(**query_set)
+
+    def host_list(self):
+        query_set = self._build_query(self._QUERY['host_list'])
+        return self._db.fetch_all_tuples(**query_set)
+
+    # TODO will be deprecated
+    def _reset_list(self, **kwargs):
+        query_set = self._build_query(self._QUERY['reset_list'], **kwargs)
+        return self._db.fetch_all_tuples(**query_set)
+
+    # TODO will be deprecated
+    def _reset_task(self, tid):
+        # Build 5 reset queries
+        queries = [
+            self._build_sql_item(self._QUERY[name], tid=tid)
+            for name in ('trash_result', 'trash_error',
+                         'delete_result', 'delete_error', 'reset_task')
+        ]
+
+        self._db.execute_multiple_dml(queries)