+++ undine/server/__main__.py
@@ -1,10 +1,10 @@
 from argparse import ArgumentParser
 from collections import namedtuple
 from datetime import datetime
-from undine.core.scheduler import TaskScheduler
-from undine.core.task import Task
-from undine.driver.driver_factory import TaskDriverFactory
-from undine.rpc.rpc_daemon import RpcDaemon
+from undine.server.scheduler import TaskScheduler
+from undine.server.task import Task
+from undine.server.driver import TaskDriverFactory
+from undine.server.rpc import RpcDaemon
 from undine.utils.exception import UndineException
 from undine.utils.system import System